feat(lib): update sort.js lib
